refactor(sync): sync_orchestrator 按域拆成 8 个 part（B2）
SyncOrchestrator 3053 行主体拆成主库 + sync_orchestrator/*.part.dart：
aggregate / collections / tombstones / books / videos / audiobooks /
dictionaries / local_audio，每个 part 是一个 `extension _SyncOrchestratorX
on SyncOrchestrator`，方法逐字搬移、零行为改动（仓库既有 reader_fushi /
video_fushi part 惯例）。

拆法由 Dart 作用域规则决定：库私有 extension 的成员在别的库不可见，所以
公开入口（syncCollections / syncDictionaries / importRemoteBooks /
syncVideoAssets / sync*Packages / syncDeletionTombstones）、run() 和
…LiveForTest 测试桥留在 class 本体，part 只放各域 _sync*Live 等私有实现。
计划里「纯转发 ForTest 清理」据此作废——它们是测试触达私有实现的唯一桥。

守卫随扫描面搬家：
- 新增 test/sync/sync_orchestrator_source_corpus.dart（复用
  helpers/part_corpus.dart：主库 + 磁盘枚举全部 part，不手写清单）；
  sync_progress_test / server_lifecycle_appmodel_guard_test 改读合并语料
  （SyncPhase.dictionaries 只剩 part 里有，单读主库会假阴）。
- interconnect_service_config_test 指向 aggregate.part.dart（主库里
  `_syncServiceConfigLive(` 只剩 run() 的调用点，从调用点往后找门控会假阴）。
- sync_audiobook_download_wiring_guard_test 指向 audiobooks.part.dart，
  写入点清单加上该 part。

验证：flutter analyze 全量零问题；flutter test test/sync test/tools
2857 过，9 红全是 update_manifest_publish_race_test 的 PowerShell 缺 bash
环境红（PATH 加 Git\bin 后 10/10 绿）。
