+++ fushi/lib/src/sync/sync_orchestrator.dart
@@ -31,6 +31,15 @@ import 'package:fushi_audio/fushi_audio.dart'
 import 'package:fushi_core/fushi_core.dart';
 import 'package:path/path.dart' as p;
 
+part 'sync_orchestrator/aggregate.part.dart';
+part 'sync_orchestrator/collections.part.dart';
+part 'sync_orchestrator/tombstones.part.dart';
+part 'sync_orchestrator/books.part.dart';
+part 'sync_orchestrator/videos.part.dart';
+part 'sync_orchestrator/audiobooks.part.dart';
+part 'sync_orchestrator/dictionaries.part.dart';
+part 'sync_orchestrator/local_audio.part.dart';
+
 /// Reserved top-level folder (under the backend root) that holds dictionary
 /// packages. It lives alongside the per-book folders, so every place that
 /// treats root children as *books* must skip it ([isReservedSyncFolderName]).
@@ -718,72 +727,6 @@ class SyncOrchestrator {
     return report;
   }
 
-  /// 云后端聚合同步：把本机统计 + 收藏经 [AggregateSyncService] 与其它设备的
-  /// per-device 快照并集合并（只增不减 / 值不缩小 / 幂等 / 并集），写回本地并
-  /// 上传本机最新快照。无 baseline / 无冲突弹窗（集合 + 单调语义无损）。首次同步
-  /// （云上无 `__aggregate__` 快照）优雅退化为「只上传本机快照」，不崩。
-  ///
-  /// 整段包在 try/catch 里，逐轮错误进 [report.errors] 不中断整体 sweep（与其它
-  /// 维度同纪律）。删除不跨端传播；无 schema 变更。
-  Future<void> _syncAggregate(SyncRunReport report) async {
-    try {
-      await AggregateSyncService(_db, scope: _scope).sync(
-        store: _backend,
-        deviceId: deviceId,
-      );
-    } catch (e) {
-      report.noteError('aggregate sync', e);
-    }
-  }
-
-  /// Pulls the host's explicitly allowlisted service configuration. The host is
-  /// authoritative for keys it publishes; no client→host write exists.
-  Future<void> _syncServiceConfigLive(
-    SyncRunReport report,
-    InterconnectSyncBackend backend,
-  ) async {
-    try {
-      // apikey 同步设定重设计：service-config（host 的外部服务 API key）此前是
-      // 无 UI 无开关的隐形通道。开关默认 true（行为不变）；关掉 = 本设备不再向
-      // host 请求 service-config（连请求都不发，不是拉回来再丢弃）。
-      if (!await SyncRepository(_db).isInterconnectServiceConfigSyncEnabled()) {
-        return;
-      }
-      final InterconnectServiceConfigSnapshot? snapshot =
-          await backend.getRemoteServiceConfig();
-      if (snapshot == null) return;
-      report.serviceConfigsImported += await snapshot.applyTo(_db);
-    } catch (e) {
-      report.noteError('service config live sync', e);
-    }
-  }
-
-  /// 互联聚合（统计 + 收藏）live 双向合并（TODO-1056 phase C）。
-  ///
-  /// 复用 [AggregateSyncService.syncOverClient] 的通道无关核心（materialize 本地 →
-  /// GET host 快照 → [AggregateMergeService] 并集折叠 → 写回本地 DB（只 MAX / 并集
-  /// upsert，幂等）→ PUT 合并后快照回 host）。IO 由本方法注入：GET 走
-  /// [InterconnectSyncBackend.getRemoteAggregate]（老 host 无端点返回 404 → null →
-  /// 只推不拉，优雅降级），PUT 走 [InterconnectSyncBackend.putRemoteAggregate]。
-  ///
-  /// 无 baseline / 无冲突弹窗（集合 + 单调语义无损）；删除不跨端传播；无 schema 变更。
-  /// 整段 try/catch，逐轮错误进 [report.errors] 不中断整体 sweep（与其它维度同纪律）。
-  Future<void> _syncAggregateLive(
-    SyncRunReport report,
-    InterconnectSyncBackend backend,
-  ) async {
-    try {
-      await AggregateSyncService(_db, scope: _scope).syncOverClient(
-        fetchRemote: backend.getRemoteAggregate,
-        pushMerged: backend.putRemoteAggregate,
-        shareStats: syncStats,
-        shareFavorites: syncFavorites,
-      );
-    } catch (e) {
-      report.noteError('aggregate live sync', e);
-    }
-  }
-
   /// 测试入口：直接调用 [_syncAggregateLive]（private 方法对测试文件不可见）。
   @visibleForTesting
   Future<void> syncAggregateLiveForTest(
@@ -966,67 +909,6 @@ class SyncOrchestrator {
     }
   }
 
-  /// 合集双向同步（多端库联合视图 §2.3 任务5.3，互联 host API 通道）。
-  ///
-  /// 读-合并-写，与云后端 [syncCollections] 完全同构，仅通道不同（host API 而非
-  /// WebDAV `__collections__` 文件箱）：GET host 合集清单
-  /// （[InterconnectSyncBackend.getRemoteCollectionManifest]）→ [CollectionSyncEngine.merge]
-  /// 与本地全量快照合并（成员并集 + 移出/删除墓碑按基线裁决 + 手动序整合集 LWW）→
-  /// [applyCollectionLocalChanges] 落本地 DB（计入 [SyncRunReport.collectionsUpdated]）→
-  /// 合并后清单与 host 字节不同才 POST 回写（host 端再经 mergeCollectionManifest 并入
-  /// 自己 DB，同一引擎）。
-  ///
-  /// 老 host 无端点（GET 404 → null）时优雅跳过（不 POST，不崩）。基线（本端
-  /// [SyncRepository.getCollectionsSyncBaselineMs]）在整轮成功后才推进；中途失败保持
-  /// 旧基线，下轮同一批墓碑重当「新闻」裁决——应用端按目标态调和幂等重放安全。
-  /// 整段 try/catch，错误进 [report.errors] 不中断整体 sweep（与其它维度同纪律）。
-  Future<void> _syncCollectionsLive(
-    SyncRunReport report,
-    InterconnectSyncBackend backend,
-  ) async {
-    try {
-      // 竞态修复：读远端清单**之前**先预取本轮基线时刻，整轮成功后写这个预取值
-      // （而非结束时的新 now），并作 publishedAt 发布时戳。
-      final int nextBaseline = DateTime.now().millisecondsSinceEpoch;
-      final CollectionManifest? remote =
-          await backend.getRemoteCollectionManifest();
-      if (remote == null) {
-        // 老 host 无 /api/library/collections 端点（对端 app 版本过旧）。以前静默
-        // return，用户只看见「合集没同步」却无任何线索（BUG-938 次因）。计入
-        // report.errors：错误日志留痕 + 手动「立即同步」按错误计数提示；其余维度
-        // 不受影响照常同步。
-        report.errors.add(
-            'collections live sync: host has no collections endpoint '
-            '(older app version) — update the host app to sync collections');
-        return;
-      }
-
-      final CollectionManifest local = await loadLocalCollectionManifest(_db);
-      final SyncRepository repo = SyncRepository(_db);
-      // 时钟回拨钳制：持久化基线晚于 now 时钳到 now。
-      int baseline = await repo.getCollectionsSyncBaselineMs(_scope);
-      if (baseline > nextBaseline) baseline = nextBaseline;
-      final CollectionSyncOutcome outcome = CollectionSyncEngine.merge(
-        local: local,
-        remote: remote,
-        lastSyncedAtMs: baseline,
-        nowMs: nextBaseline,
-      );
-
-      report.collectionsUpdated +=
-          await applyCollectionLocalChanges(_db, outcome.changes);
-
-      // 回写门槛：字节有变才 POST（确定性排序保证内容相等 ⇒ 字节相等，避免每轮
-      // 无谓写放大）。
-      if (outcome.merged.canonicalJson() != remote.canonicalJson()) {
-        await backend.putRemoteCollectionManifest(outcome.merged);
-      }
-      await repo.setCollectionsSyncBaselineMs(_scope, nextBaseline);
-    } catch (e) {
-      report.noteError('collections live sync', e);
-    }
-  }
-
   /// 测试入口：直接调用 [_syncCollectionsLive]（private 方法对测试文件不可见）。
   @visibleForTesting
   Future<void> syncCollectionsLiveForTest(
@@ -1034,64 +916,6 @@ class SyncOrchestrator {
     InterconnectSyncBackend backend,
   ) =>
       _syncCollectionsLive(report, backend);
-
-  /// 收集本设备当前在库的资产（按 mediaType 分组 → `itemKey → 存在起始时刻`），供删除
-  /// 墓碑消费端算 deleteLocal 候选（远端有删除标记 ∧ 本地仍在库 ∧ 标记管得着这条）。
-  /// itemKey 与写墓碑点严格一致：book/audiobook = bookKey（[writeSyncDeletionTombstone]
-  /// 调用点 reader_fushi_source / audiobook），video = bookUid（video_book_repository），
-  /// localaudio = displayName，srtbook = srt_books.uid（仅 standalone，见下）。
-  ///
-  /// 键一律用 [SyncTombstoneKind.dbValue] 而非裸字符串字面量：这张 map 与写墓碑点必须
-  /// 逐字一致，拼错一个字符的后果是「对端删了、本地永远不弹确认」这种静默失效。
-  ///
-  /// BUG-2044：值是**存在起始时刻**而不再是单纯的「在不在库」。删除后又重新加回来的条目
-  /// 时刻晚于墓碑 deletedAt，[tombstoneAppliesTo] 据此把它排除在候选之外——否则本机自己
-  /// 取消收藏产生、发布到远端后再也不会被 GC 的那条墓碑，会在用户重新收藏同一句之后被读
-  /// 回来，弹「其他设备已删除」问用户要不要删掉自己刚收藏的东西。
-  Future<DeletionPresentEntries> _collectPresentDeletionKeys() async {
-    return <String, Map<String, int?>>{
-      SyncTombstoneKind.book.dbValue: <String, int?>{
-        for (final EpubBookRow r in await _db.getAllEpubBooks())
-          r.bookKey: r.importedAt,
-      },
-      // 有声书表没有自己的导入时刻列，且与 epub 共享 bookKey——借 epub 的 importedAt
-      // 会把「书早就在、有声书是后加的」记成前者。宁可 null（多问一次），不编造时刻。
-      SyncTombstoneKind.audiobook.dbValue: <String, int?>{
-        for (final AudiobookRow r in await _db.getAllAudiobooks())
-          r.bookKey: null,
-      },
-      // 纯字幕书（standalone SRT）身份 = uid。**只收 bookKey 为空的行**：与
-      // [SrtBookRepository.delete] 的写墓碑判据同源——srt-backed 行的身份是 bookKey，
-      // 已由上面的 book 键覆盖，重复收进来会让同一资产在对端弹两条确认（TODO-2470）。
-      SyncTombstoneKind.srtbook.dbValue: <String, int?>{
-        for (final SrtBookRow r in await _db.getAllSrtBooks())
-          if (r.bookKey.isEmpty) r.uid: r.importedAt,
-      },
-      SyncTombstoneKind.video.dbValue: <String, int?>{
-        for (final VideoBookRow r in await _db.allVideoBooks())
-          r.bookUid: r.importedAt,
-      },
-      // localaudio 条目不记录加入时刻 → null = 无从仲裁，保持「只看在不在库」的旧语义。
-      SyncTombstoneKind.localaudio.dbValue: <String, int?>{
-        for (final LocalAudioDbEntry e in localAudioEntries)
-          e.displayName: null,
-      },
-      SyncTombstoneKind.favoriteword.dbValue: <String, int?>{
-        for (final FavoriteWordRow r in await _db.getAllFavoriteWords())
-          FushiDatabase.favoriteWordItemKey(
-              r.expression, r.reading, r.sourceType): r.createdAt,
-      },
-      // 收藏句无稳定 id，用内容键（[FavoriteSentenceRepository.itemKeyOf]）；与写墓碑点、
-      // aggregate 去重键同源。时刻取 createdAt——重新收藏会生成新的 createdAt，正是
-      // BUG-2044 仲裁所依据的那个时刻。
-      SyncTombstoneKind.favoritesentence.dbValue: <String, int?>{
-        for (final FavoriteSentence s
-            in await FavoriteSentenceRepository(_db).getAll())
-          FavoriteSentenceRepository.itemKeyOf(s):
-              s.createdAt.millisecondsSinceEpoch,
-      },
-    };
-  }
 
   /// 删除墓碑同步（云后端通道，显式确认式删除传播 Phase C/D）。
   ///
@@ -1219,184 +1043,6 @@ class SyncOrchestrator {
       }
     } catch (e) {
       report.noteError('deletion tombstones sync', e);
-    }
-  }
-
-  /// 删除墓碑同步（互联 host API 通道），**双向**：
-  ///
-  /// 1. 推送（client→host，[_pushDeletionTombstonesLive]）：把本机「从所有设备删除」
-  ///    产生的墓碑真的删到对端 host 上。host 自己的 delete 会写它自己的墓碑，于是
-  ///    第三台设备下轮照常收到确认提示——链路闭合。
-  /// 2. 消费（host→client）：GET host 墓碑（老 host 404 → null 优雅跳过）→ 与本地在库键
-  ///    求交 deleteLocal 候选 → 过基线守卫 → 塞 report，UI 弹逐条确认。
-  ///
-  /// 此处曾是 GET-only（注释原文「client 自身删除不经此推给 host，各端自行确认删除」），
-  /// 后果是勾了「从所有设备删除」对互联对端完全无效——墓碑只留在本地表里没人发布。
-  /// 消费语义仍与云 [syncDeletionTombstones] 一致；推送是互联独有（云通道的对应动作是
-  /// 往 `__tombstones__` 写标记，两者各记各的基线，见
-  /// [SyncRepository.getDeletionTombstonesPushBaselineMs]）。
-  Future<void> _syncDeletionTombstonesLive(
-    SyncRunReport report,
-    InterconnectSyncBackend backend,
-  ) async {
-    try {
-      final int nextBaseline = DateTime.now().millisecondsSinceEpoch;
-      // 先推后拉：本机的删除意图先发出去，再看对端有什么要删的。两步共用同一个
-      // nextBaseline（预取，防 IO 期间时钟漂移造成窗口空洞）。推送整段自带 try，
-      // 失败不挡消费。
-      await _pushDeletionTombstonesLive(report, backend, nextBaseline);
-
-      final List<({String mediaType, String itemKey, int deletedAt})>? remote =
-          await backend.getRemoteDeletionTombstones();
-      if (remote == null) return; // 老 host 无 /api/tombstones 端点，优雅跳过。
-
-      final DeletionTombstoneEntries remoteTombstones =
-          <String, Map<String, int>>{};
-      for (final r in remote) {
-        final Map<String, int> byKey =
-            remoteTombstones.putIfAbsent(r.mediaType, () => <String, int>{});
-        final int? prev = byKey[r.itemKey];
-        if (prev == null || r.deletedAt > prev) {
-          byKey[r.itemKey] = r.deletedAt;
-        }
-      }
-
-      final SyncRepository repo = SyncRepository(_db);
-      final DeletionPresentEntries present =
-          await _collectPresentDeletionKeys();
-      int baseline = await repo.getDeletionTombstonesBaselineMs(_scope);
-      if (baseline > nextBaseline) baseline = nextBaseline;
-
-      final List<DeletionPropagationCandidate> raw = computeDeletionPropagation(
-        localTombstones: const <String, Map<String, int>>{},
-        remoteTombstones: remoteTombstones,
-        localPresent: present,
-        remotePresent: const <String, Map<String, int?>>{},
-      );
-      for (final DeletionPropagationCandidate c in raw) {
-        if (c.direction != DeletionPropagationDirection.deleteLocal) continue;
-        final int? at = remoteTombstones[c.mediaType]?[c.itemKey];
-        if (at == null || at <= baseline) continue;
-        report.deletionCandidates.add(c);
-        report.noteDeletionHighWater(_scope, at);
-      }
-    } catch (e) {
-      report.noteError('deletion tombstones live sync', e);
-    }
-  }
-
-  /// 把本机未推送的删除墓碑推给对端 host（client→host，「从所有设备删除」的落地端）。
-  ///
-  /// 墓碑只在用户显式选 [DeleteScope.syncEverywhere] 时才写（各实体删除路径的门控），
-  /// 所以走到这里的每一条都是用户明确要求「所有设备都删」的条目——推送不需要再问一次。
-  ///
-  /// 因果轴用 [SyncRepository.getDeletionTombstonesPushBaselineMs]：只推
-  /// `baseline < deletedAt <= nextBaseline` 的墓碑，整批成功才推进基线。**不碰墓碑行上的
-  /// `remotePublishedAt`**——那是云通道 `__tombstones__` 的账，互联去标它会让同时配了云
-  /// 备份的设备永远跳过这条、只连云的第三台设备再也收不到这次删除。
-  ///
-  /// 两类失败区别对待：
-  /// * **异常**（网络 / host 5xx）→ 不推进基线，下轮整批重试。这类失败通常是整体性的
-  ///   （断网），整批重试正是想要的；DELETE 端点幂等，重推已成功的无害。
-  /// * **host 不支持**（视频 DELETE 端点 404/405 = 对端版本过旧）→ 记 `report.errors`
-  ///   但**不**阻塞基线。能力缺失不是暂时性故障，为它永久卡住基线会让书 / 有声书的
-  ///   删除每轮无谓重推。代价是对端升级前的这条视频删除会漏掉——UI 侧长按删除路径会
-  ///   直接提示「对端版本过旧」，同步路径则留在日志里。
-  Future<void> _pushDeletionTombstonesLive(
-    SyncRunReport report,
-    InterconnectSyncBackend backend,
-    int nextBaseline,
-  ) async {
-    try {
-      final SyncRepository repo = SyncRepository(_db);
-      int baseline = await repo.getDeletionTombstonesPushBaselineMs();
-      // 时钟回拨钳制：基线晚于本轮取的 now 时按 now 算，否则一次回拨会把窗口永久关死。
-      if (baseline > nextBaseline) baseline = nextBaseline;
-
-      final List<SyncDeletionTombstoneRow> rows =
-          await _db.getSyncDeletionTombstones();
-      bool retryable = false;
-      for (final SyncDeletionTombstoneRow row in rows) {
-        // 窗口两端都要卡：晚于 nextBaseline 的（本地时钟超前写出的未来戳）留到下轮，
-        // 否则推进基线会把它一并盖掉、这条删除就此蒸发。
-        if (row.deletedAt <= baseline || row.deletedAt > nextBaseline) continue;
-        final SyncTombstoneKind? kind =
-            SyncTombstoneKind.tryParse(row.mediaType);
-        // 未知 kind = 比本端新的版本写的墓碑，本端不认识就别猜着删（前向兼容）。
-        if (kind == null) continue;
-        if (!_hasInterconnectDeletionChannel(kind)) continue;
-        try {
-          final bool supported =
-              await _pushOneDeletionLive(backend, kind, row.itemKey);
-          if (!supported) {
-            report.errors.add(
-              'host does not support deleting ${row.mediaType} '
-              '"${row.itemKey}" (peer app too old); skipped',
-            );
-          }
-        } catch (e) {
-          retryable = true;
-          report.noteError(
-            'deletion push ${row.mediaType}/${row.itemKey}',
-            e,
-          );
-        }
-      }
-      if (!retryable) {
-        await repo.setDeletionTombstonesPushBaselineMs(nextBaseline);
-      }
-    } catch (e) {
-      report.noteError('deletion tombstones live push', e);
-    }
-  }
-
-  /// 该 kind 在互联 live 通道上有没有删除端点。
-  ///
-  /// 收藏词 / 收藏句没有：它们经聚合快照通道同步，而 `applyAggregateSnapshot` 按设计
-  /// 只做 MAX / 并集（删除不跨端传播）。这里如实跳过——不假装推过（那会静默丢删除），
-  /// 也不算作失败（那会为一件永远做不成的事永久卡住基线）。
-  static bool _hasInterconnectDeletionChannel(SyncTombstoneKind kind) {
-    switch (kind) {
-      case SyncTombstoneKind.book:
-      case SyncTombstoneKind.audiobook:
-      case SyncTombstoneKind.srtbook:
-      case SyncTombstoneKind.localaudio:
-      case SyncTombstoneKind.video:
-        return true;
-      case SyncTombstoneKind.favoriteword:
-      case SyncTombstoneKind.favoritesentence:
-        return false;
-    }
-  }
-
-  /// 按 kind 分派到对应的 host 删除端点。返回 host 是否支持该删除（false = 对端版本
-  /// 过旧，端点 404/405）；其余失败照常抛给调用方计入 retryable。
-  ///
-  /// 纯字幕书（srtbook）与有声书共用 `DELETE /api/library/audiobooks/<identity>`：host
-  /// 端 identity 解析同时查 `Audiobooks(bookKey)` 与 `SrtBooks(uid)`，两种键都能命中。
-  Future<bool> _pushOneDeletionLive(
-    InterconnectSyncBackend backend,
-    SyncTombstoneKind kind,
-    String itemKey,
-  ) async {
-    switch (kind) {
-      case SyncTombstoneKind.book:
-        await backend.deleteRemoteBook(itemKey);
-        return true;
-      case SyncTombstoneKind.audiobook:
-      case SyncTombstoneKind.srtbook:
-        await backend.deleteRemoteAudiobook(itemKey);
-        return true;
-      case SyncTombstoneKind.localaudio:
-        await backend.deleteRemoteLocalAudio(itemKey);
-        return true;
-      case SyncTombstoneKind.video:
-        // 唯一会如实报「不支持」的一条：视频删除端点是本次新增的，旧 host 没有。
-        return backend.deleteRemoteVideo(itemKey);
-      case SyncTombstoneKind.favoriteword:
-      case SyncTombstoneKind.favoritesentence:
-        // [_hasInterconnectDeletionChannel] 已挡在前面，走不到这里。
-        return true;
     }
   }
 
@@ -1568,23 +1214,6 @@ class SyncOrchestrator {
     }
   }
 
-  /// 一条 `VideoBooks` 行是否是可作为单文件上传的**本地**视频（[syncVideoAssets] 用）。
-  ///
-  /// 排除：流媒体（`streamSpecJson` 非空 / `videoPath` 为 http(s) URL）——无本地字节
-  /// 可传；多集播放列表（`playlistJson` 非空）——单文件资产模型装不下多集（多集上传
-  /// 是后续批的接缝，见 §2.6 seam）。
-  bool _isUploadableLocalVideo(VideoBookRow v) {
-    if (v.streamSpecJson != null) return false;
-    if (v.playlistJson != null) return false;
-    final String path = v.videoPath;
-    if (path.isEmpty) return false;
-    final String lower = path.toLowerCase();
-    if (lower.startsWith('http://') || lower.startsWith('https://')) {
-      return false;
-    }
-    return true;
-  }
-
   /// Folds the per-book sweep results into [SyncRunReport.conflicts]. Only
   /// [SyncResult.conflict] rows are collected; everything else (imported /
   /// exported / synced / skipped) is left to the existing per-phase tallies
@@ -1651,544 +1280,6 @@ class SyncOrchestrator {
         }
       } catch (e) {
         report.noteError('import book "${folder.name}"', e);
-      }
-    }
-  }
-
-  /// 互联书籍内容 live 上传。
-  ///
-  /// 直打对端 `/api/library/books` 端点，按 `sanitizeTtuFilename(title)` 只处理
-  /// toPush：本端有 && 远端无 → `repackageExtractedEpub` 重打包 →
-  /// `putRemoteBook` 上传。远端独有书籍留给 compare/interconnect UI 手动下载。
-  ///
-  /// 仅当 client syncContent 开时由 [run] 调用。进度走 [SyncPhase.books]，
-  /// 临时文件 finally 清理，逐项错误进 [report.errors] 不中断整体。
-  ///
-  /// **删除传播**：现有实现不传播书籍删除（SyncManager 云路径同语义）。
-  /// 若后续需要互联书籍删除传播，参考词典删除传播（BUG-086）扩展此方法。
-  Future<void> _syncBooksContentLive(
-    SyncRunReport report,
-    InterconnectSyncBackend backend,
-  ) async {
-    final List<RemoteBookInfo> remoteBooks = await backend.listRemoteBooks();
-    final List<EpubBookRow> localBooks = await _db.getAllEpubBooks();
-
-    final Set<String> localKeys = <String>{
-      for (final EpubBookRow b in localBooks) sanitizeTtuFilename(b.title),
-    };
-    final Map<String, bool> remoteKeyHasContent = <String, bool>{
-      // 「远端已有内容」= EPUB 内容树 ∨ 漫画包内容（互联完整支持批次）——否则
-      // host 上已有的漫画会被本端当「远端无」反复推送。
-      for (final RemoteBookInfo r in remoteBooks)
-        sanitizeTtuFilename(r.title): r.hasContent || r.hasMangaContent,
-    };
-
-    // 按 sanitizeTtuFilename(title) union 计算 diff。
-    final BookSyncDiff diff = computeBookSyncDiff(
-      localKeys: localKeys,
-      remoteKeyHasContent: remoteKeyHasContent,
-    );
-
-    // 需要本地 title 原始值用于端点调用（端点按原始 title 寻址）。
-    final Map<String, String> localKeyToTitle = <String, String>{
-      for (final EpubBookRow b in localBooks)
-        sanitizeTtuFilename(b.title): b.title,
-    };
-
-    // BUG-1503：本机用户给这些书改的名字 + LWW 戳，随上传一起走 header。一趟读
-    // 完（推多本书只查一次偏好表），没改过名的书查不到 → 不发 header。
-    final Map<String, OverrideTitleEntry> overrideTitles =
-        await readOverrideTitlesByBookKey(_db);
-
-    final int total = diff.toPush.length;
-    int index = 0;
-
-    // ── Push：本端独有 → 重打包并上传 ───────────────────────────────────────
-    for (final String key in diff.toPush) {
-      final String title = localKeyToTitle[key] ?? key;
-      _emit(SyncPhase.books, itemIndex: index, itemTotal: total, title: title);
-      File? tmp;
-      try {
-        // 找到本地行取 extractDir。
-        final EpubBookRow? row = localBooks.cast<EpubBookRow?>().firstWhere(
-              (EpubBookRow? b) => sanitizeTtuFilename(b!.title) == key,
-              orElse: () => null,
-            );
-        if (row == null ||
-            row.extractDir.isEmpty ||
-            !Directory(row.extractDir).existsSync()) {
-          // 本地内容不可用，跳过（与 importRemoteBooks 对称语义）。
-          report.errors
-              .add('live push book "$title": extractDir missing or empty');
-          index++;
-          continue;
-        }
-        final BookFormat format = BookFormat.parseOrEpub(row.format);
-        if (format == BookFormat.pdf) {
-          // PDF 无互联内容通道（互联全域盘点已记录）。此前无 format 过滤时每本
-          // 漫画/PDF 每轮同步都稳定产出一条 repackage 失败噪音错误——静默跳过。
-          index++;
-          continue;
-        }
-        tmp = _tmpFile('.epub');
-        // 漫画 → 书目录整树 zip（manga.json 标记，host importBook 内容嗅探分流）；
-        // EPUB → 既有 repackage。
-        final bool built = format == BookFormat.manga
-            ? await repackageMangaBook(row.extractDir, tmp.path)
-            : await repackageExtractedEpub(row.extractDir, tmp.path);
-        if (!built) {
-          report.errors
-              .add('live push book "$title": repackage produced no output');
-          index++;
-          continue;
-        }
-        // 显示名跟着书走，**身份不跟着走**（BUG-1488 定的红线）：端点寻址、
-        // host 端 bookKey 派生仍恒用 raw [title]。
-        final OverrideTitleEntry? override = overrideTitles[row.bookKey];
-        await backend.putRemoteBook(
-          title,
-          tmp,
-          displayTitle: override?.title,
-          displayTitleAt: override?.updatedAt ?? 0,
-          onProgress: (double f) => _emit(SyncPhase.books,
-              itemIndex: index,
-              itemTotal: total,
-              title: title,
-              fileFraction: f),
-        );
-      } catch (e) {
-        report.noteError('live push book "$title"', e);
-      } finally {
-        _safeDelete(tmp);
-      }
-      index++;
-    }
-  }
-
-  /// 互联书籍阅读进度 live 双向同步（TODO-767）。
-  ///
-  /// 遍历本地 `epub_books`，对每本书：GET host 真相源进度（[RemoteBookClient
-  /// .remoteBookProgress]，host 直读自己的 `reader_positions`）+ 读本地
-  /// `reader_positions`，用 [resolveBookProgressSync]「取较新时间戳」选胜者；胜者
-  /// 严格新于 host 时 PUT 上报 host（[RemoteBookClient.putRemoteBookProgress]，host
-  /// 再防御性取较新落自己的 DB），胜者不同于本地时 upsert 回本地。
-  ///
-  /// 修复根因：互联「立即同步」此前书籍进度只走 SyncManager 的 WebDAV 文件箱
-  /// （progress_*.json），host 从不读回自己的 reader_positions DB，故进度不过去。
-  /// 这里补对称视频 TODO-653 的 live 端点 + host-apply，让进度真正落 host DB。
-  ///
-  /// 逐本错误进 [report.errors] 不中断整体。
-  Future<void> _syncBookProgressLive(
-    SyncRunReport report,
-    InterconnectSyncBackend backend,
-  ) async {
-    final List<EpubBookRow> localBooks = await _db.getAllEpubBooks();
-    for (final EpubBookRow book in localBooks) {
-      try {
-        final RemoteBookProgress remote =
-            await backend.remoteBookProgress(book.bookKey);
-        // v82：wire 键仍是 bookKey（REST 路径冻结），本地子表键是书 uid。
-        final ReaderPositionRow? localRow =
-            await _db.getReaderPosition(book.uid);
-        final RemoteBookProgress local = localRow == null
-            ? RemoteBookProgress.empty
-            : RemoteBookProgress(
-                sectionIndex: localRow.sectionIndex,
-                normCharOffset: localRow.normCharOffset,
-                charOffset: localRow.charOffset,
-                updatedAtMs: localRow.updatedAt,
-              );
-        final RemoteBookProgress winner =
-            resolveBookProgressSync(local: local, remote: remote);
-
-        // 本地→host：胜者严格新于 host 时上报（host 端再取较新，幂等安全）。
-        if (winner.updatedAtMs > remote.updatedAtMs ||
-            (winner.updatedAtMs == remote.updatedAtMs &&
-                (winner.sectionIndex != remote.sectionIndex ||
-                    winner.normCharOffset != remote.normCharOffset ||
-                    winner.charOffset != remote.charOffset))) {
-          await backend.putRemoteBookProgress(book.bookKey, winner);
-        }
-
-        // host→本地：胜者不同于本地时 upsert 回本地 reader_positions。
-        final bool localChanged = winner.sectionIndex != local.sectionIndex ||
-            winner.normCharOffset != local.normCharOffset ||
-            winner.charOffset != local.charOffset ||
-            winner.updatedAtMs != local.updatedAtMs;
-        if (localChanged && winner.updatedAtMs > 0) {
-          await _db.upsertReaderPosition(ReaderPositionsCompanion(
-            bookUid: Value(book.uid),
-            sectionIndex: Value(winner.sectionIndex),
-            normCharOffset: Value(winner.normCharOffset),
-            charOffset: Value(winner.charOffset),
-            updatedAt: Value(winner.updatedAtMs),
-          ));
-          // BUG-686: a host-newer progress pull lands in reader_positions but
-          // writes no book content, so it must still flag the shelf for a
-          // refresh — the cached fushiBooksProvider otherwise keeps showing the
-          // pre-sync progress bar and the sync looks like it did nothing.
-          report.localBookProgressPulled++;
-        }
-      } catch (e) {
-        report.noteError('live book progress "${book.title}"', e);
-      }
-    }
-  }
-
-  /// 互联播放断点 live 双向 sweep 的共享模板（视频 / 有声书）。
-  ///
-  /// 两条 sweep 历史上 ~90% 逐字同构，仅四个探针不同，命名统一轮收口于此。
-  /// 对 [localKeys] ∩ [hostKeys] 里的每个键：
-  /// 1. [readLocal] 取本地 (位置, 时间戳)，[readHost] 取 host (位置, 时间戳)；
-  /// 2. [resolvePositionLww]「取较新时间戳」选胜者；
-  /// 3. 本地→host：胜者新于 host（或同戳不同位）→ [pushToHost] 上报
-  ///    （host 端再取较新，幂等安全）；
-  /// 4. host→本地：胜者不同于本地 → [writeBackLocal] 写回。
-  ///
-  /// 只对 host 也有的键同步（本地独有条目无 host 真相，跳过）；逐条错误以
-  /// `[errorLabel] "<key>": <e>` 进 [report.errors] 不中断整体。host 清单的获取
-  /// 与两侧空集合的早退仍在各调用方（保持既有网络行为不变）。
-  Future<void> _syncPositionsLive(
-    SyncRunReport report, {
-    required String errorLabel,
-    required Set<String> localKeys,
-    required Set<String> hostKeys,
-    required Future<({int positionMs, int updatedAtMs})> Function(String key)
-        readLocal,
-    required Future<({int positionMs, int updatedAtMs})> Function(String key)
-        readHost,
-    required Future<void> Function(String key, int positionMs, int updatedAtMs)
-        pushToHost,
-    required Future<void> Function(String key, int positionMs, int updatedAtMs)
-        writeBackLocal,
-  }) async {
-    for (final String key in localKeys) {
-      if (!hostKeys.contains(key)) continue; // host 无此条目：跳过。
-      try {
-        final ({int positionMs, int updatedAtMs}) local = await readLocal(key);
-        final ({int positionMs, int updatedAtMs}) host = await readHost(key);
-
-        final ({int positionMs, int updatedAtMs}) winner = resolvePositionLww(
-          localPositionMs: local.positionMs,
-          localUpdatedAtMs: local.updatedAtMs,
-          remotePositionMs: host.positionMs,
-          remoteUpdatedAtMs: host.updatedAtMs,
-        );
-
-        // 本地→host：胜者新于 host 时上报（host 端再取较新，幂等安全）。
-        if (winner.updatedAtMs > host.updatedAtMs ||
-            (winner.updatedAtMs == host.updatedAtMs &&
-                winner.positionMs != host.positionMs)) {
-          await pushToHost(key, winner.positionMs, winner.updatedAtMs);
-        }
-
-        // host→本地：胜者不同于本地时写回。
-        if (winner.positionMs != local.positionMs ||
-            winner.updatedAtMs != local.updatedAtMs) {
-          await writeBackLocal(key, winner.positionMs, winner.updatedAtMs);
-        }
-      } catch (e) {
-        report.noteError('$errorLabel "$key"', e);
-      }
-    }
-  }
-
-  /// 互联视频播放进度 live 双向同步（TODO-767 + TODO-816）。
-  ///
-  /// 此前只遍历本地 `VideoBooks`，对每条比对 host 进度。但 client 流式看的远端视频
-  /// 在本地**没有 VideoBooks 行**（[home_video_page._openRemote] 只 push 不 upsert，
-  /// 进度只落 `video_remote_position_<uid>` prefs，见 video_fushi_page._persistRemotePosition）
-  /// → 旧 sweep 永远扫不到，流式视频进度无法纳入全量双向同步（TODO-816 子问题1 断点①）。
-  ///
-  /// 修复：同步基底统一为 uid 集合 = 「本地 VideoBooks 行 uid」∪「本地有
-  /// `video_remote_position_<uid>` prefs 的 uid（哪怕无行）」，只对 host 也有的 uid 同步。
-  ///
-  /// 本地进度真相：书架视频与流式视频共用一个 `_at_` prefs 时间戳。书架视频的位置
-  /// 在 `VideoBooks.lastPositionMs`（本机播放写），流式视频的位置在
-  /// `video_remote_position_<uid>` prefs（resume 路径写）——同一进度的两处镜像，不会
-  /// 同时各有不同含义。本地位置取「有行用 lastPositionMs，否则用 prefs 位置」，时间戳
-  /// 统一取 `_at_` prefs（无则 0=旧数据，被任何带时间戳的远端进度盖过）。
-  ///
-  /// 写回时位置真相统一落 `video_remote_position_<uid>` + `_at_` prefs（resume 路径
-  /// 同键空间）；**仅当本地存在 VideoBooks 行时**才一并更新 `lastPositionMs`
-  /// （流式视频绝不强建行污染书架）。
-  ///
-  /// LWW 比对与逐条错误处理见共享模板 [_syncPositionsLive]。
-  Future<void> _syncVideoProgressLive(
-    SyncRunReport report,
-    InterconnectSyncBackend backend,
-  ) async {
-    // 基底①：本地 VideoBooks 行（uid → lastPositionMs，向后兼容书架视频）。
-    final Map<String, int> rowPositionByUid = <String, int>{};
-    for (final VideoBookRow video in await _db.allVideoBooks()) {
-      rowPositionByUid[video.bookUid] = video.lastPositionMs;
-    }
-    // 基底②：本地看过的流式视频（有 video_remote_position_<uid> prefs 但无行）。
-    final Set<String> localUids = <String>{...rowPositionByUid.keys};
-    final Map<String, String> allPrefs = await _db.getAllPrefs();
-    for (final String key in allPrefs.keys) {
-      final String? uid = videoUidFromRemotePositionPrefKey(key);
-      if (uid != null) localUids.add(uid);
-    }
-    if (localUids.isEmpty) return;
-
-    // 视频进度是 host-truth 模型：进度端点只对 host DB 里真实存在的视频可用。先取
-    // host 视频清单（条目已带 positionMs / positionUpdatedAtMs，省去逐视频 GET），
-    // 只对两端都有的视频同步；本地独有视频（host 无）无 host 真相可同步，跳过。
-    final Map<String, RemoteVideoInfo> hostById = <String, RemoteVideoInfo>{};
-    for (final RemoteVideoInfo info in await backend.listRemoteVideos()) {
-      hostById[info.id] = info;
-    }
-    if (hostById.isEmpty) return;
-
-    await _syncPositionsLive(
-      report,
-      errorLabel: 'live video progress',
-      localKeys: localUids,
-      hostKeys: hostById.keys.toSet(),
-      readLocal: (String uid) async {
-        final int prefsPos =
-            await _db.getPrefTyped<int>(videoRemotePositionPrefKey(uid), 0);
-        final int prefsAt =
-            await _db.getPrefTyped<int>(videoRemotePositionAtPrefKey(uid), 0);
-        return (
-          positionMs: rowPositionByUid[uid] ?? prefsPos,
-          updatedAtMs: prefsAt,
-        );
-      },
-      // host 清单条目已带 positionMs / positionUpdatedAtMs，无需逐视频 GET。
-      readHost: (String uid) async {
-        final RemoteVideoInfo info = hostById[uid]!;
-        return (
-          positionMs: info.positionMs,
-          updatedAtMs: info.positionUpdatedAtMs,
-        );
-      },
-      pushToHost: (String uid, int positionMs, int updatedAtMs) =>
-          backend.putRemoteVideoPosition(uid, positionMs, updatedAtMs),
-      writeBackLocal: (String uid, int positionMs, int updatedAtMs) async {
-        await _db.setPrefTyped<int>(
-            videoRemotePositionPrefKey(uid), positionMs);
-        await _db.setPrefTyped<int>(
-            videoRemotePositionAtPrefKey(uid), updatedAtMs);
-        if (rowPositionByUid.containsKey(uid)) {
-          // 时刻用**对端的** updatedAtMs（不是 now）：这是「对方在那个时刻看到这
-          // 里」的事实。传 now 会把三天前的远端进度冒充成本机刚看的，直接把合集
-          // 续播锚点（BUG-1542）钉在一集用户根本没在看的集上。
-          await _db.updateVideoBookPosition(uid, positionMs,
-              playedAt: updatedAtMs);
-        }
-      },
-    );
-
-    // 播放偏好双向收敛（BUG-1620 调轴起步，播放偏好同步泛化批扩展为全字段：
-    // 调轴/音轨/副字幕源/副字幕调轴）。与进度共用同一 uid 基底 + host 清单带戳
-    // 字段（零额外网络读）；逐字段「严格较新时间戳者胜」。本地较新字段聚合成一次
-    // putRemoteVideoPlayback（旧 host 无端点 404 → 静默吞，不刷 report 噪音）；
-    // host 较新字段写回本地键对（时间戳用对端的，同进度写回纪律）+ 本地有
-    // VideoBooks 行时写穿行值，本机播放立即跟随。
-    String? nullableStr(String raw) => raw.isEmpty ? null : raw;
-    for (final String uid in localUids) {
-      final RemoteVideoInfo? info = hostById[uid];
-      if (info == null) continue;
-      try {
-        final VideoPlaybackSyncState host = info.playback;
-        final VideoPlaybackSyncState local = VideoPlaybackSyncState(
-          delayMs: await _db.getPrefTyped<int>(videoRemoteDelayPrefKey(uid), 0),
-          delayAt:
-              await _db.getPrefTyped<int>(videoRemoteDelayAtPrefKey(uid), 0),
-          audioTrackId: nullableStr(await _db.getPrefTyped<String>(
-              videoRemoteAudioTrackPrefKey(uid), '')),
-          audioTrackAt: await _db.getPrefTyped<int>(
-              videoRemoteAudioTrackAtPrefKey(uid), 0),
-          secondarySubtitleSource: nullableStr(await _db.getPrefTyped<String>(
-              videoRemoteSecondarySubtitlePrefKey(uid), '')),
-          secondarySubtitleAt: await _db.getPrefTyped<int>(
-              videoRemoteSecondarySubtitleAtPrefKey(uid), 0),
-          secondaryDelayMs: int.tryParse(await _db.getPrefTyped<String>(
-              videoRemoteSecondaryDelayPrefKey(uid), '')),
-          secondaryDelayAt: await _db.getPrefTyped<int>(
-              videoRemoteSecondaryDelayAtPrefKey(uid), 0),
-        );
-
-        // 本地→host：只带本地严格较新的字段（at 置 0 的字段 host 端 merge 忽略）。
-        final VideoPlaybackSyncState push = VideoPlaybackSyncState(
-          delayMs: local.delayMs,
-          delayAt: local.delayAt > host.delayAt ? local.delayAt : 0,
-          audioTrackId: local.audioTrackId,
-          audioTrackAt:
-              local.audioTrackAt > host.audioTrackAt ? local.audioTrackAt : 0,
-          secondarySubtitleSource: local.secondarySubtitleSource,
-          secondarySubtitleAt:
-              local.secondarySubtitleAt > host.secondarySubtitleAt
-                  ? local.secondarySubtitleAt
-                  : 0,
-          secondaryDelayMs: local.secondaryDelayMs,
-          secondaryDelayAt: local.secondaryDelayAt > host.secondaryDelayAt
-              ? local.secondaryDelayAt
-              : 0,
-        );
-        if (!push.isEmpty) {
-          try {
-            await backend.putRemoteVideoPlayback(uid, push);
-          } catch (e) {
-            debugPrint(
-                '[SyncOrchestrator] video playback push "$uid" failed: $e');
-          }
-        }
-
-        // host→本地：逐字段写回严格较新者。
-        final bool hasRow = rowPositionByUid.containsKey(uid);
-        final VideoPlaybackSyncState merged =
-            VideoPlaybackSyncState.merge(local, host);
-        if (merged.delayAt != local.delayAt) {
-          await _db.setPrefTyped<int>(
-              videoRemoteDelayPrefKey(uid), merged.delayMs);
-          await _db.setPrefTyped<int>(
-              videoRemoteDelayAtPrefKey(uid), merged.delayAt);
-          if (hasRow) await _db.updateVideoBookDelayMs(uid, merged.delayMs);
-        }
-        if (merged.audioTrackAt != local.audioTrackAt) {
-          await _db.setPrefTyped<String>(
-              videoRemoteAudioTrackPrefKey(uid), merged.audioTrackId ?? '');
-          await _db.setPrefTyped<int>(
-              videoRemoteAudioTrackAtPrefKey(uid), merged.audioTrackAt);
-          if (hasRow) {
-            await _db.updateVideoBookAudioTrackId(uid, merged.audioTrackId);
-          }
-        }
-        if (merged.secondarySubtitleAt != local.secondarySubtitleAt) {
-          await _db.setPrefTyped<String>(
-              videoRemoteSecondarySubtitlePrefKey(uid),
-              merged.secondarySubtitleSource ?? '');
-          await _db.setPrefTyped<int>(
-              videoRemoteSecondarySubtitleAtPrefKey(uid),
-              merged.secondarySubtitleAt);
-          if (hasRow) {
-            await _db.updateVideoBookSecondarySubtitleSource(
-                uid, merged.secondarySubtitleSource);
-          }
-        }
-        if (merged.secondaryDelayAt != local.secondaryDelayAt) {
-          await _db.setPrefTyped<String>(videoRemoteSecondaryDelayPrefKey(uid),
-              merged.secondaryDelayMs?.toString() ?? '');
-          await _db.setPrefTyped<int>(
-              videoRemoteSecondaryDelayAtPrefKey(uid), merged.secondaryDelayAt);
-          if (hasRow) {
-            await _db.updateVideoBookSecondaryDelayMs(
-                uid, merged.secondaryDelayMs);
-          }
-        }
-      } catch (e) {
-        report.noteError('live video playback "$uid"', e);
-      }
-    }
-  }
-
-  /// 互联有声书播放进度 live 双向同步（BUG-471）。
-  ///
-  /// 与视频 [_syncVideoProgressLive] 完全对称（共享模板 [_syncPositionsLive]）。
-  /// client 听的远端有声书进度落 `audiobook_pos_<bookKey>` +
-  /// `audiobook_pos_at_<bookKey>` prefs（见 [AudiobookRepository.updatePositionMs]），
-  /// 但互联角色非对称：host 只跑 server，从不回灌自己的 `audiobook_pos_` pref，
-  /// 故「立即同步」点了进度不过去（云后端经 SyncManager 双向文件箱正常，互联缺这一段）。
-  ///
-  /// 同步基底 = 「本地有 `audiobook_pos_<key>` prefs 的 bookKey」∪「本地 Audiobooks
-  /// 行的 bookKey」，只对 host 也有的 bookKey 同步（host 无该有声书时其 PUT 端点
-  /// 404 / 闸门 no-op，且 GET 无真相可拉，跳过省一次网络）。本地进度真相 =
-  /// `audiobook_pos_<bookKey>`（位置）+ `audiobook_pos_at_<bookKey>`（时间戳，
-  /// 旧数据无记 0）；写回落同一 prefs 键空间（同 resume/播放写）。
-  Future<void> _syncAudiobookProgressLive(
-    SyncRunReport report,
-    InterconnectSyncBackend backend,
-  ) async {
-    // 同步基底：本地 Audiobooks 行 ∪ 本地有 audiobook_pos_<key> prefs 的 bookKey。
-    final Set<String> localKeys = <String>{
-      for (final AudiobookRow ab in await _db.getAllAudiobooks()) ab.bookKey,
-    };
-    final Map<String, String> allPrefs = await _db.getAllPrefs();
-    for (final String key in allPrefs.keys) {
-      final String? bookKey = audiobookKeyFromPositionPrefKey(key);
-      if (bookKey != null) localKeys.add(bookKey);
-    }
-    if (localKeys.isEmpty) return;
-
-    // 有声书进度是 host-truth 模型：只对 host 也有的有声书同步。先取 host 有声书
-    // 清单，只同步两端都有的键；本地独有有声书无 host 真相，跳过。
-    //
-    // 键必须用 [RemoteAudiobookInfo.identity]（srt-backed=bookKey；纯 SRT=uid），
-    // 不能用裸 bookKey：纯 SRT standalone 的 bookKey 恒空串，用它建交集会让这类书
-    // 的 `audiobook_pos_<uid>` 永远落不进 hostKeys → 听书进度跨设备完全不同步——
-    // 而 host 端 get/putAudiobookPosition 本就按 identity（bookKey ∪ SrtBooks.uid）
-    // 命中（BUG-1637）。空 identity（异常行）跳过不进集合。
-    final Map<String, RemoteAudiobookInfo> hostById =
-        <String, RemoteAudiobookInfo>{
-      for (final RemoteAudiobookInfo info
-          in await backend.listRemoteAudiobooks())
-        if (info.identity.isNotEmpty) info.identity: info,
-    };
-    if (hostById.isEmpty) return;
-
-    await _syncPositionsLive(
-      report,
-      errorLabel: 'live audiobook progress',
-      localKeys: localKeys,
-      hostKeys: hostById.keys.toSet(),
-      readLocal: (String bookKey) async => (
-        positionMs:
-            await _db.getPrefTyped<int>(audiobookPositionPrefKey(bookKey), 0),
-        updatedAtMs:
-            await _db.getPrefTyped<int>(audiobookPositionAtPrefKey(bookKey), 0),
-      ),
-      // 新 host 清单已内联断点（互联完整支持批次）→ 免逐本 GET；旧 host 清单无此
-      // 字段（0@0）→ 退回逐本 GET（真无进度的书多打一次也无害，幂等）。
-      readHost: (String bookKey) async {
-        final RemoteAudiobookInfo info = hostById[bookKey]!;
-        if (info.positionUpdatedAtMs > 0) {
-          return (
-            positionMs: info.positionMs,
-            updatedAtMs: info.positionUpdatedAtMs,
-          );
-        }
-        return backend.remoteAudiobookPosition(bookKey);
-      },
-      pushToHost: (String bookKey, int positionMs, int updatedAtMs) =>
-          backend.putRemoteAudiobookPosition(bookKey, positionMs, updatedAtMs),
-      writeBackLocal: (String bookKey, int positionMs, int updatedAtMs) async {
-        await _db.setPrefTyped<int>(
-            audiobookPositionPrefKey(bookKey), positionMs);
-        await _db.setPrefTyped<int>(
-            audiobookPositionAtPrefKey(bookKey), updatedAtMs);
-      },
-    );
-
-    // 有声书调轴双向收敛（互联完整支持批次；与视频 delay sweep 同范式）。
-    // 「严格较新时间戳者胜」；两侧都无戳（旧数据/从未调过）无事可做。旧 host 无
-    // /delay 端点 → push 404 静默吞（best-effort，不刷 report 噪音）。
-    for (final String identity in localKeys) {
-      final RemoteAudiobookInfo? info = hostById[identity];
-      if (info == null) continue;
-      try {
-        final int localDelay =
-            await _db.getPrefTyped<int>(audiobookDelayPrefKey(identity), 0);
-        final int localAt =
-            await _db.getPrefTyped<int>(audiobookDelayAtPrefKey(identity), 0);
-        if (localAt > info.delayUpdatedAtMs) {
-          try {
-            await backend.putRemoteAudiobookDelay(
-                identity, localDelay, localAt);
-          } catch (e) {
-            debugPrint(
-                '[SyncOrchestrator] audiobook delay push "$identity" failed: $e');
-          }
-        } else if (info.delayUpdatedAtMs > localAt) {
-          await _db.setPrefTyped<int>(
-              audiobookDelayPrefKey(identity), info.delayMs);
-          await _db.setPrefTyped<int>(
-              audiobookDelayAtPrefKey(identity), info.delayUpdatedAtMs);
-        }
-      } catch (e) {
-        report.noteError('live audiobook delay "$identity"', e);
       }
     }
   }
@@ -2268,572 +1359,6 @@ class SyncOrchestrator {
       return;
     }
     await syncLocalAudioPackages(report, direction: direction);
-  }
-
-  /// 互联直读对端实时词典：按名 union，绝不创建/读写 __dictionaries__。
-  Future<void> _syncDictionariesLive(
-    SyncRunReport report,
-    InterconnectSyncBackend backend,
-    SyncAssetDirection direction,
-  ) async {
-    final List<DictionaryMetaRow> localDicts =
-        await _db.getAllDictionaryMetadata();
-    final List<RemoteDictionaryInfo> remoteDicts =
-        await backend.listRemoteDictionaries();
-
-    final SyncKeyDiff diff = computeKeyUnionDiff(
-      localKeys: <String>{for (final DictionaryMetaRow d in localDicts) d.name},
-      remoteKeys: <String>{
-        for (final RemoteDictionaryInfo d in remoteDicts) d.name
-      },
-    );
-
-    // 方向裁剪放在**循环之外**：total 从一开始就是这次真正要做的量，循环体一行不
-    // 改。把 if 塞进循环里只会让进度分母撒谎（显示 0/5 却只做 2 件事）。
-    final List<String> pulls = <String>[if (direction.pulls) ...diff.toPull];
-    final List<String> pushes = <String>[if (direction.pushes) ...diff.toPush];
-    final int total = pulls.length + pushes.length;
-    int index = 0;
-
-    for (final String name in pulls) {
-      _emit(SyncPhase.dictionaries,
-          itemIndex: index, itemTotal: total, title: name);
-      File? tmp;
-      try {
-        tmp = _tmpFile(_dictionaryAssetSuffix);
-        await backend.getRemoteDictionary(name, tmp,
-            onProgress: (double f) => _emit(SyncPhase.dictionaries,
-                itemIndex: index,
-                itemTotal: total,
-                title: name,
-                fileFraction: f));
-        await _packages.importDictionaryPackage(
-          packageFile: tmp,
-          dictionaryResourceRoot: _dictionaryResourceRoot,
-        );
-        report.dictionariesImported++;
-      } catch (e) {
-        report.noteError('pull dictionary "$name"', e);
-      } finally {
-        _safeDelete(tmp);
-      }
-      index++;
-    }
-
-    for (final String name in pushes) {
-      _emit(SyncPhase.dictionaries,
-          itemIndex: index, itemTotal: total, title: name);
-      File? tmp;
-      try {
-        tmp = _tmpFile(_dictionaryAssetSuffix);
-        await _packages.exportDictionaryPackage(
-          dictionaryName: name,
-          dictionaryResourceRoot: _dictionaryResourceRoot,
-          outputFile: tmp,
-        );
-        await backend.putRemoteDictionary(name, tmp,
-            onProgress: (double f) => _emit(SyncPhase.dictionaries,
-                itemIndex: index,
-                itemTotal: total,
-                title: name,
-                fileFraction: f));
-        report.dictionariesExported++;
-      } catch (e) {
-        report.noteError('push dictionary "$name"', e);
-      } finally {
-        _safeDelete(tmp);
-      }
-      index++;
-    }
-  }
-
-  /// Union-syncs dictionary packages in the `__dictionaries__` namespace.
-  Future<void> _syncDictionariesStaged(
-    SyncRunReport report,
-    SyncAssetDirection direction,
-  ) async {
-    final String ns = await _backend.ensureNamespace(kSyncDictionaryNamespace);
-    final List<DictionaryMetaRow> localDicts =
-        await _db.getAllDictionaryMetadata();
-    final List<AssetEntry> remote = await _backend.listChildren(ns);
-
-    final Set<String> remoteNames = <String>{
-      for (final AssetEntry e in remote)
-        if (!e.isFolder && _isDictionaryAsset(e.name))
-          _stripDictionaryAssetSuffix(e.name),
-    };
-    final Set<String> localNames = <String>{
-      for (final DictionaryMetaRow d in localDicts) d.name,
-    };
-
-    // Resolve both sides' work first so progress has a real denominator.
-    final List<DictionaryMetaRow> toPush = <DictionaryMetaRow>[
-      if (direction.pushes)
-        for (final DictionaryMetaRow d in localDicts)
-          if (!remoteNames.contains(d.name)) d,
-    ];
-    final List<AssetEntry> toPull = <AssetEntry>[
-      if (direction.pulls)
-        for (final AssetEntry e in remote)
-          if (!e.isFolder &&
-              _isDictionaryAsset(e.name) &&
-              !localNames.contains(_stripDictionaryAssetSuffix(e.name)))
-            e,
-    ];
-    final int total = toPush.length + toPull.length;
-    int index = 0;
-
-    // Push local-only dictionaries.
-    for (final DictionaryMetaRow d in toPush) {
-      _emit(SyncPhase.dictionaries,
-          itemIndex: index, itemTotal: total, title: d.name);
-      File? tmp;
-      try {
-        tmp = _tmpFile(_dictionaryAssetSuffix);
-        await _packages.exportDictionaryPackage(
-          dictionaryName: d.name,
-          dictionaryResourceRoot: _dictionaryResourceRoot,
-          outputFile: tmp,
-        );
-        await _backend.putAsset(ns, '${d.name}$_dictionaryAssetSuffix', tmp,
-            onProgress: (double f) => _emit(SyncPhase.dictionaries,
-                itemIndex: index,
-                itemTotal: total,
-                title: d.name,
-                fileFraction: f));
-        report.dictionariesExported++;
-      } catch (e) {
-        report.noteError('export dictionary "${d.name}"', e);
-      } finally {
-        _safeDelete(tmp);
-      }
-      index++;
-    }
-
-    // Pull remote-only dictionaries.
-    for (final AssetEntry e in toPull) {
-      // Show the clean dictionary name in progress, matching the push side —
-      // the asset name still carries the `.fushidict`（或 Hibiki 时代的
-      // `.fushidict`）suffix, which otherwise surfaces as a "weird" entry in
-      // the progress list.
-      final String displayName = _stripDictionaryAssetSuffix(e.name);
-      _emit(SyncPhase.dictionaries,
-          itemIndex: index, itemTotal: total, title: displayName);
-      File? tmp;
-      try {
-        tmp = _tmpFile(_dictionaryAssetSuffix);
-        await _backend.getAsset(e.id, tmp,
-            onProgress: (double f) => _emit(SyncPhase.dictionaries,
-                itemIndex: index,
-                itemTotal: total,
-                title: displayName,
-                fileFraction: f));
-        await _packages.importDictionaryPackage(
-          packageFile: tmp,
-          dictionaryResourceRoot: _dictionaryResourceRoot,
-        );
-        report.dictionariesImported++;
-      } catch (err) {
-        report.noteError('import dictionary "${e.name}"', err);
-      } finally {
-        _safeDelete(tmp);
-      }
-      index++;
-    }
-  }
-
-  /// 互联本地音频 live 同步（Phase 3 T3.4）。
-  ///
-  /// 直打对端 `/api/library/localaudio` 端点，按 `displayName` union：
-  /// - toPull：远端有 ∧ 本端无 → `getRemoteLocalAudio` 下载包 → `onLocalAudioImported` 注册；
-  /// - toPush：本端有 ∧ 远端无 → `exportLocalAudioPackage` 打包 → `putRemoteLocalAudio` 上传。
-  ///
-  /// [direction] 裁剪要做哪一半（显式上传 / 下载动作用；[SyncAssetDirection.both]
-  /// 保留完整 union 语义）。只由 [syncLocalAudioSources] 分派调用，不再随 [run] 跑。
-  /// 进度走 [SyncPhase.localAudio]，临时文件 finally 清理，逐项错误进 report.errors 不中断。
-  Future<void> _syncLocalAudioLive(
-    SyncRunReport report,
-    InterconnectSyncBackend backend,
-    SyncAssetDirection direction,
-  ) async {
-    final List<RemoteLocalAudioInfo> remoteEntries =
-        await backend.listRemoteLocalAudio();
-    final Set<String> localNames = <String>{
-      for (final LocalAudioDbEntry d in localAudioEntries) d.displayName,
-    };
-    final Set<String> remoteNames = <String>{
-      for (final RemoteLocalAudioInfo r in remoteEntries) r.displayName,
-    };
-
-    final SyncKeyDiff diff = computeKeyUnionDiff(
-      localKeys: localNames,
-      remoteKeys: remoteNames,
-    );
-
-    // 方向裁剪在循环之外（同 [_syncDictionariesLive]）：total 即本次真实工作量。
-    final List<String> pulls = <String>[if (direction.pulls) ...diff.toPull];
-    final List<String> pushes = <String>[if (direction.pushes) ...diff.toPush];
-    final int total = pulls.length + pushes.length;
-    int index = 0;
-
-    // ── Pull：远端独有 → 下载并注册 ────────────────────────────────────────
-    for (final String name in pulls) {
-      _emit(SyncPhase.localAudio,
-          itemIndex: index, itemTotal: total, title: name);
-      File? tmp;
-      File? stagingDb;
-      try {
-        tmp = _tmpFile(_localAudioAssetSuffix);
-        await backend.getRemoteLocalAudio(
-          name,
-          tmp,
-          onProgress: (double f) => _emit(SyncPhase.localAudio,
-              itemIndex: index, itemTotal: total, title: name, fileFraction: f),
-        );
-        final LocalAudioPackageContents contents =
-            await _packages.importLocalAudioPackage(
-          packageFile: tmp,
-          stagingDir: _tempDir,
-        );
-        stagingDb = contents.dbFile;
-        if (onLocalAudioImported != null) {
-          await onLocalAudioImported!(contents);
-          report.localAudioImported++;
-        }
-      } catch (e) {
-        report.noteError('live pull local audio "$name"', e);
-      } finally {
-        _safeDelete(tmp);
-        _safeDelete(stagingDb);
-      }
-      index++;
-    }
-
-    // ── Push：本端独有 → 打包并上传 ─────────────────────────────────────────
-    for (final String name in pushes) {
-      _emit(SyncPhase.localAudio,
-          itemIndex: index, itemTotal: total, title: name);
-      File? tmp;
-      try {
-        final LocalAudioDbEntry? entry =
-            localAudioEntries.cast<LocalAudioDbEntry?>().firstWhere(
-                  (LocalAudioDbEntry? d) => d!.displayName == name,
-                  orElse: () => null,
-                );
-        if (entry == null || !File(entry.path).existsSync()) {
-          report.errors.add(
-              'live push local audio "$name": DB file missing or not found');
-          index++;
-          continue;
-        }
-        tmp = _tmpFile(_localAudioAssetSuffix);
-        await _packages.exportLocalAudioPackage(
-          displayName: entry.displayName,
-          enabled: entry.enabled,
-          sources: entry.sources,
-          dbFile: File(entry.path),
-          outputFile: tmp,
-        );
-        await backend.putRemoteLocalAudio(
-          name,
-          tmp,
-          onProgress: (double f) => _emit(SyncPhase.localAudio,
-              itemIndex: index, itemTotal: total, title: name, fileFraction: f),
-        );
-        report.localAudioExported++;
-      } catch (e) {
-        report.noteError('live push local audio "$name"', e);
-      } finally {
-        _safeDelete(tmp);
-      }
-      index++;
-    }
-  }
-
-  /// 互联有声书包 live 双向同步（TODO-809：立即/自动同步双向拉取）。
-  ///
-  /// 直打对端 `/api/library/audiobooks` 端点，按 `bookKey` union：
-  /// - Push（本端有 ∧ 远端无）→ `exportAudioDatabasePackage` 打包 → `putRemoteAudiobook`。
-  /// - Pull（远端有 ∧ 本端无有声书）→ `getRemoteAudiobook` 下载 →
-  ///   `importAudioDatabasePackage` 解包落盘。
-  ///
-  /// **Pull 防孤儿约束**：`importAudioDatabasePackage` 只 upsert Audiobooks/SrtBooks
-  /// 行，不创建 EpubBooks 行。故只对「本端已有同 bookKey 的 EPUB、但当前缺音频」的
-  /// 远端项拉取——否则会落下没有书可绑的孤儿有声书行（这正是历史上选 push-only 的
-  /// 动机）。无对应本地 EPUB 的远端有声书跳过并记一条 info 级 error，留给手动下载
-  /// （书架远端书卡 / 同步对比对话框）补音频。拉取时用本地 EPUB 的 bookKey 作
-  /// `bookKeyOverride`，保证写入行与本地 EPUB 字节相等可配对（徽章亮）。
-  ///
-  /// 仅当 client syncAudioBookFiles 开且 isInterconnect 时由 [run] 调用。
-  /// 进度走 [SyncPhase.audiobooks]，临时文件 finally 清理，逐项错误进 report.errors 不中断。
-  Future<void> _syncAudiobooksLive(
-    SyncRunReport report,
-    InterconnectSyncBackend backend,
-  ) async {
-    final List<RemoteAudiobookInfo> remoteAudiobooks =
-        await backend.listRemoteAudiobooks();
-    final List<AudiobookRow> localAudiobooks = await _db.getAllAudiobooks();
-    final List<EpubBookRow> localBooks = await _db.getAllEpubBooks();
-
-    final Set<String> localKeys = <String>{
-      for (final AudiobookRow ab in localAudiobooks) ab.bookKey,
-    };
-    // 纯 SRT（standalone）远端有声书（bookKey 空、身份=uid）**不进自动 union**：与
-    // 远端独有 EPUB 一样是「手动下载才落地」（TODO-1291 / 书架远端占位卡），自动
-    // sweep 只处理 srt-backed（bookKey 非空）有声书文件补拉，避免把独有 standalone
-    // 书自动灌进对端，也避免空 bookKey 污染 diff。
-    final Set<String> remoteKeys = <String>{
-      for (final RemoteAudiobookInfo r in remoteAudiobooks)
-        if (r.bookKey.isNotEmpty) r.bookKey,
-    };
-    // 本端已有 EPUB 的 bookKey 集合：Pull 只对「本端有书但缺音频」的远端项动作，
-    // 避免落下无 EpubBooks 行可绑的孤儿有声书（importAudioDatabasePackage 不建书行）。
-    final Set<String> localBookKeys = <String>{
-      for (final EpubBookRow b in localBooks) b.bookKey,
-    };
-
-    final SyncKeyDiff diff = computeKeyUnionDiff(
-      localKeys: localKeys,
-      remoteKeys: remoteKeys,
-    );
-
-    // toPullAudioOnly（场景B）= 远端有 ∧ 本端无有声书 ∧ 本端已有同 bookKey
-    // EPUB → 只补音频不重导 EPUB。这是「同步有声书文件」开关下**唯一**的 pull
-    // 语义：只对本端已在库的书补/拉其音频文件。
-    //
-    // TODO-1291（用户决策 A · 解耦）：远端独有（本端完全没有这本书）的书
-    // **不**在自动同步里导入/灌书架，即便远端书带有声书且 hasContent。这类书回归
-    // 手动下载入口（compare 对比页 / 书架远端卡），与本类文档契约（见类头
-    // :126-130「remote-only EPUBs stay remote until the user explicitly
-    // downloads them」）一致。历史 TODO-873 的 toPullFullBook 自动灌书路径已在此
-    // 摘除——它把「开启同步有声书文件」误当成「自动把远端独有书拉进书架」。
-    final List<String> toPullAudioOnly = <String>[
-      for (final String key in diff.toPull)
-        if (localBookKeys.contains(key)) key,
-    ];
-
-    final int total = diff.toPush.length + toPullAudioOnly.length;
-    int index = 0;
-
-    // ── Push：本端独有 → 打包并上传 ─────────────────────────────────────────
-    for (final String key in diff.toPush) {
-      _emit(SyncPhase.audiobooks,
-          itemIndex: index, itemTotal: total, title: key);
-      File? tmp;
-      try {
-        final SrtBookRow? srt = await _db.getSrtBookByBookKey(key);
-        if (srt == null) {
-          report.errors
-              .add('live push audiobook "$key": srtBook not found, skipping');
-          index++;
-          continue;
-        }
-        tmp = _tmpFile('.fushiaudio');
-        await _packages.exportAudioDatabasePackage(
-          bookKey: key,
-          srtBookUid: srt.uid,
-          outputFile: tmp,
-        );
-        await backend.putRemoteAudiobook(
-          key,
-          tmp,
-          onProgress: (double f) => _emit(SyncPhase.audiobooks,
-              itemIndex: index, itemTotal: total, title: key, fileFraction: f),
-        );
-        report.audiobooksExported++;
-      } catch (e) {
-        report.noteError('live push audiobook "$key"', e);
-      } finally {
-        _safeDelete(tmp);
-      }
-      index++;
-    }
-
-    // ── Pull A（场景B）：远端有、本端有书但缺音频 → 下载并解包落盘 ────────────
-    for (final String key in toPullAudioOnly) {
-      _emit(SyncPhase.audiobooks,
-          itemIndex: index, itemTotal: total, title: key);
-      File? tmp;
-      try {
-        tmp = _tmpFile('.fushiaudio');
-        await backend.getRemoteAudiobook(
-          key,
-          tmp,
-          onProgress: (double f) => _emit(SyncPhase.audiobooks,
-              itemIndex: index, itemTotal: total, title: key, fileFraction: f),
-        );
-        // 用本地 EPUB 的 bookKey 作 override：远端 key 已等于本地 EPUB 的 bookKey
-        // （toPull 已由 localBookKeys 筛过），显式 override 保写入行与 EPUB 可配对。
-        await _packages.importAudioDatabasePackage(
-          packageFile: tmp,
-          audioDatabaseRoot: _audioDatabaseRoot,
-          bookKeyOverride: key,
-        );
-        report.audiobooksImported++;
-      } catch (e) {
-        report.noteError('live pull audiobook "$key"', e);
-      } finally {
-        _safeDelete(tmp);
-      }
-      index++;
-    }
-  }
-
-  /// 互联视频文件 live push（client→host，TODO §2.6「后续批」接线）。
-  ///
-  /// 直打对端 host 上传端点：枚举本地可上传单文件视频（[_isUploadableLocalVideo]，
-  /// 排除流媒体 / 多集播放列表），对 host 尚无（按 bookUid）或尺寸不同的推上去。视频
-  /// 身份键是 `VideoBooks.bookUid`（= host 端 [RemoteVideoInfo.id]，两端同源派生），故
-  /// 直接按 uid union，重复上传同一视频 host 端 upsert 覆盖同一行、不产生重复。
-  ///
-  /// **upload-only**：host→client 方向仍是按需流式播放 / 手动下载（视频 GB 级不进
-  /// 自动 pull），与云后端 [syncVideoAssets] 同律。互联 host 上传字节未混淆（与
-  /// [putRemoteAudiobook] 同为裸 octet-stream），故 host 清单 [RemoteVideoInfo.sizeBytes]
-  /// == 本地明文尺寸，可直接按尺寸判幂等（不必像云后端那样绕物理尺寸走清单）。
-  ///
-  /// 仅当 client syncVideoFiles 开且 isInterconnect 时由 [run] 调用。进度走
-  /// [SyncPhase.videos]，逐项错误进 report.errors 不中断，[report.videosExported] 计上传数。
-  Future<void> _syncVideosLive(
-    SyncRunReport report,
-    InterconnectSyncBackend backend,
-  ) async {
-    // host 既有视频（uid → 清单条目；sizeBytes null=host 无法 stat）。
-    final Map<String, RemoteVideoInfo> hostByUid = <String, RemoteVideoInfo>{
-      for (final RemoteVideoInfo info in await backend.listRemoteVideos())
-        info.id: info,
-    };
-
-    // 稳定顺序（uid 升序）遍历本地可上传视频，先算出真正要传的（host 无 / 尺寸不同），
-    // 让进度分母只计实际上传数。字幕（BUG-964）：视频本轮上传 ⇒ 其全部 sidecar 一并
-    // 推；host 已有视频但清单报无外挂字幕 ⇒ 只补推字幕（不重传视频）。
-    final List<VideoBookRow> localVideos = <VideoBookRow>[
-      for (final VideoBookRow v in await _db.allVideoBooks())
-        if (_isUploadableLocalVideo(v)) v,
-    ]..sort((VideoBookRow a, VideoBookRow b) => a.bookUid.compareTo(b.bookUid));
-
-    final Map<String, List<String>?> sidecarDirCache =
-        <String, List<String>?>{};
-    final List<({VideoBookRow row, File file, bool pushVideo})> toPush =
-        <({VideoBookRow row, File file, bool pushVideo})>[];
-    for (final VideoBookRow v in localVideos) {
-      final File file = File(v.videoPath);
-      if (!file.existsSync()) {
-        report.errors.add(
-            'live push video "${v.title}": local file missing: ${v.videoPath}');
-        continue;
-      }
-      final RemoteVideoInfo? host = hostByUid[v.bookUid];
-      final int? hostSize = host?.sizeBytes;
-      final int localSize = await file.length();
-      // host 已有：尺寸可比且相等 ⇒ 跳过（幂等）；host 尺寸不可知（null）也跳过（无从
-      // 判差异，避免每轮全量重传大视频，下轮 host stat 成功即自愈）。host 无 ⇒ 上传。
-      final bool pushVideo =
-          !(host != null && (hostSize == null || hostSize == localSize));
-      // 字幕补推判据与视频同粒度：host 已有任一 sidecar 即跳过（改字幕内容/后加语言
-      // 不重推，与视频同尺寸跳过一致）。
-      // pushVideo==false 蕴含 host!=null（流程分析已提升，无需判空）。
-      final bool pushSubtitles = (pushVideo || !host.hasSubtitle) &&
-          _localSidecarSubtitles(v.videoPath, sidecarDirCache).isNotEmpty;
-      if (!pushVideo && !pushSubtitles) continue;
-      toPush.add((row: v, file: file, pushVideo: pushVideo));
-    }
-
-    final int total = toPush.length;
-    int index = 0;
-    // 老 host 无字幕端点（首个 404/405）后停止本轮后续字幕推送，只记一条可见提示
-    // （与合集端点缺失的可见性纪律一致），不把每条视频都刷成一条错误。
-    bool subtitleEndpointMissing = false;
-    for (final ({VideoBookRow row, File file, bool pushVideo}) item in toPush) {
-      final VideoBookRow v = item.row;
-      _emit(SyncPhase.videos,
-          itemIndex: index, itemTotal: total, title: v.title);
-      bool videoOk = !item.pushVideo;
-      if (item.pushVideo) {
-        try {
-          await backend.putRemoteVideo(
-            v.bookUid,
-            item.file,
-            title: v.title,
-            onProgress: (double f) => _emit(SyncPhase.videos,
-                itemIndex: index,
-                itemTotal: total,
-                title: v.title,
-                fileFraction: f),
-          );
-          report.videosExported++;
-          videoOk = true;
-        } catch (e) {
-          report.noteError('live push video "${v.title}"', e);
-        }
-      }
-      // 字幕跟着视频走：视频本轮失败就不推字幕（host 侧无行可挂）。
-      if (videoOk && !subtitleEndpointMissing) {
-        subtitleEndpointMissing = !await _pushVideoSubtitlesLive(
-          report: report,
-          backend: backend,
-          row: v,
-          sidecarDirCache: sidecarDirCache,
-        );
-        if (subtitleEndpointMissing) {
-          report.errors.add(
-              'live push subtitles: host has no subtitle endpoint (older app '
-              'version) — update the host app to sync video subtitles');
-        }
-      }
-      index++;
-    }
-  }
-
-  /// 把 [row] 视频的全部本地 sidecar 字幕推给 host（BUG-964）。
-  ///
-  /// 返回 false 表示 host 无字幕端点（老版本，404/405），调用方停止本轮后续字幕
-  /// 推送；单条字幕的其它失败进 [report.errors] 不中断。
-  Future<bool> _pushVideoSubtitlesLive({
-    required SyncRunReport report,
-    required InterconnectSyncBackend backend,
-    required VideoBookRow row,
-    required Map<String, List<String>?> sidecarDirCache,
-  }) async {
-    final String stem = p.basenameWithoutExtension(row.videoPath);
-    for (final File sub
-        in _localSidecarSubtitles(row.videoPath, sidecarDirCache)) {
-      final String suffix = p.basename(sub.path).substring(stem.length);
-      try {
-        final bool supported = await backend
-            .putRemoteVideoSubtitle(row.bookUid, sub, suffix: suffix);
-        if (!supported) return false;
-      } catch (e) {
-        report.noteError('live push subtitle "${p.basename(sub.path)}"', e);
-      }
-    }
-    return true;
-  }
-
-  /// 列出 [videoPath] 同目录属于它的全部 sidecar 字幕文件（[listSidecarSubtitles]
-  /// 匹配规则）。[sidecarDirCache] 按目录缓存一次 listSync（同目录多视频的 sweep
-  /// 不重复扫盘）；目录不可读缓存 null → 返回空。
-  List<File> _localSidecarSubtitles(
-    String videoPath,
-    Map<String, List<String>?> sidecarDirCache,
-  ) {
-    final String dir = p.dirname(videoPath);
-    final List<String>? names = sidecarDirCache.putIfAbsent(dir, () {
-      try {
-        return Directory(dir)
-            .listSync(followLinks: false)
-            .whereType<File>()
-            .map((FileSystemEntity f) => p.basename(f.path))
-            .toList();
-      } on FileSystemException {
-        return null;
-      }
-    });
-    if (names == null) return const <File>[];
-    return <File>[
-      for (final String name
-          in listSidecarSubtitles(p.basenameWithoutExtension(videoPath), names))
-        File(p.join(dir, name)),
-    ];
   }
 
   /// Union-syncs local audio source DBs in the `__local_audio__` namespace.
